% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expected_mbias.R
\name{calc_expected_mbias}
\alias{calc_expected_mbias}
\title{Function to calculate multiplicative bias (observed / actual) in variant prevalence}
\usage{
calc_expected_mbias(p_v1, c_ratio)
}
\arguments{
\item{p_v1}{actual variant prevalence (proportion)}

\item{c_ratio}{coefficient of detection ratio, calculated as the ratio of the coefficients of variant 1 to variant 2}
}
\value{
scalar giving the multiplicative bias of variant 1
}
\description{
This function calculates the multiplicative bias of the observed variant proportion in \eqn{H_v1} relative to the actual variant proportion in \eqn{N_v1}.
This function assumes that variant 1 is the variant of concern (VOC).
This function is specific to the two-variant system.
}
\examples{
calc_expected_mbias(p_v1 = 0.1, c_ratio = 1.1)

}
\seealso{
Other VOC_bias: 
\code{\link{calc_observed_freq}()}
}
\author{
Shirlee Wohl, Elizabeth C. Lee, Bethany L. DiPrete, and Justin Lessler
}
\concept{VOC_bias}
