% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_freq_logistic.R
\name{calc_freq_logistic}
\alias{calc_freq_logistic}
\title{Function to calculate observed variant prevalence at time t given logistic growth (PDF)}
\usage{
calc_freq_logistic(t, p0, r, c_ratio = 1)
}
\arguments{
\item{t}{time step number (e.g., days) at which to calculate prevalence}

\item{r}{logistic growth rate}

\item{c_ratio}{coefficient of detection ratio, calculated as the ratio of the coefficients of variant 1 to variant 2; default = 1 (no bias)}

\item{p.0}{initial variant prevalence (# introductions / population size)}
}
\value{
scalar giving the variant prevalence at time t
}
\description{
Function to calculate observed variant prevalence at time t given logistic growth (PDF)
}
\examples{
calc_freq_logistic(t = 30, p0 = 1/10000, r = 0.1, c_ratio = 1)

}
\seealso{
Other VOC_loggrowth: 
\code{\link{calc_cdf_logistic}()}
}
\author{
Shirlee Wohl, Elizabeth C. Lee, Bethany L. DiPrete, and Justin Lessler
}
\concept{VOC_loggrowth}
