% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_observed_freq.R
\name{calc_observed_freq}
\alias{calc_observed_freq}
\title{Function to calculate observed variant prevalence}
\usage{
calc_observed_freq(p_v1, c_ratio)
}
\arguments{
\item{p_v1}{actual variant prevalence (proportion)}

\item{c_ratio}{coefficient of detection ratio, calculated as the ratio of the coefficients of variant 1 to variant 2}
}
\value{
scalar of observed prevalence of variant 1
}
\description{
This function calculates the observed variant prevalence in \eqn{H_v1} from the coefficient of detection ratio and the actual variant prevalence \eqn{N_v1}.
This function assumes that variant 1 is the variant of concern (VOC).
This function is specific to the two-variant system.
}
\examples{
calc_observed_freq(p_v1 = 0.1, c_ratio = 1.1)

}
\seealso{
Other VOC_bias: 
\code{\link{calc_expected_mbias}()}
}
\author{
Shirlee Wohl, Elizabeth C. Lee, Bethany L. DiPrete, and Justin Lessler
}
\concept{VOC_bias}
