% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prob_detect_cont.R
\name{calc_prob_detect_cont}
\alias{calc_prob_detect_cont}
\title{Function to calculate probability of detecting a variant given a per-timestep sample size assuming periodic sampling}
\usage{
calc_prob_detect_cont(n, t, p0, r, c_ratio = 1)
}
\arguments{
\item{n}{per-timestep (e.g., per day) sample size}

\item{t}{time step number (e.g., days) at which variant should be detected by}

\item{p0}{initial variant prevalence (# introductions / population size)}

\item{r}{logistic growth rate}

\item{c_ratio}{coefficient of detection ratio, calculated as the ratio of the coefficients of variant 1 to variant 2. default = 1 (no bias)}
}
\value{
scalar of detection probability
}
\description{
Function to calculate probability of detecting a variant given a per-timestep sample size assuming periodic sampling
}
\examples{
calc_prob_detect_cont(n = 158, t = 30, p0 = 1/10000, r = 0.1, c_ratio = 1)

}
\seealso{
Other VOC_detection: 
\code{\link{calc_samplesize_detect_cont}()},
\code{\link{calc_samplesize_detect}()}
}
\author{
Shirlee Wohl, Elizabeth C. Lee, Bethany L. DiPrete, and Justin Lessler
}
\concept{VOC_detection}
