% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_samplesize_detect.R
\name{calc_samplesize_detect}
\alias{calc_samplesize_detect}
\title{Function to calculate sample size needed for variant detection assuming cross-sectional sampling}
\usage{
calc_samplesize_detect(p_v1, prob, c_ratio)
}
\arguments{
\item{p_v1}{variant prevalence (proportion)}

\item{prob}{desired probability of detection}

\item{c_ratio}{coefficient of detection ratio, calculated as the ratio of the coefficients of variant 1 to variant 2}
}
\value{
scalar of expected sample size
}
\description{
Function to calculate sample size needed for variant detection assuming cross-sectional sampling
}
\examples{
calc_samplesize_detect(p_v1 = 0.1, prob = 0.95, c_ratio = 1)

}
\seealso{
Other VOC_detection: 
\code{\link{calc_prob_detect_cont}()},
\code{\link{calc_samplesize_detect_cont}()}
}
\author{
Shirlee Wohl, Elizabeth C. Lee, Bethany L. DiPrete, and Justin Lessler
}
\concept{VOC_detection}
