% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_samplesize_detect_cont.R
\name{calc_samplesize_detect_cont}
\alias{calc_samplesize_detect_cont}
\title{Function to calculate sample size needed for variant detection assuming periodic sampling}
\usage{
calc_samplesize_detect_cont(prob, t, p0, r, c_ratio = 1)
}
\arguments{
\item{prob}{desired probability of detection}

\item{t}{time step number (e.g., days) at which variant should be detected by}

\item{p0}{initial variant prevalence (# introductions / population size)}

\item{r}{logistic growth rate}

\item{c_ratio}{coefficient of detection ratio, calculated as the ratio of the coefficients of variant 1 to variant 2. default = 1 (no bias)}
}
\value{
scalar of expected sample size
}
\description{
Function to calculate sample size needed for variant detection assuming periodic sampling
}
\examples{
calc_samplesize_detect_cont(prob = 0.95, t = 30, p0 = 1/10000, r = 0.1, c_ratio = 1)

}
\seealso{
Other VOC_detection: 
\code{\link{calc_prob_detect_cont}()},
\code{\link{calc_samplesize_detect}()}
}
\author{
Shirlee Wohl, Elizabeth C. Lee, Bethany L. DiPrete, and Justin Lessler
}
\concept{VOC_detection}
