% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_samplesize_prev.R
\name{calc_samplesize_prev}
\alias{calc_samplesize_prev}
\title{Function to calculate sample size needed for variant prevalence estimation under cross-sectional sampling}
\usage{
calc_samplesize_prev(p_v1, prob, precision, c_ratio)
}
\arguments{
\item{p_v1}{variant prevalence (proportion)}

\item{prob}{desired confidence in variant prevalence estimate}

\item{precision}{desired precision in variant prevalence estimate}

\item{c_ratio}{coefficient of detection ratio, calculated as the ratio of the coefficients of variant 1 to variant 2. default = 1 (no bias)}
}
\value{
scalar of sample size
}
\description{
Function to calculate sample size needed for variant prevalence estimation under cross-sectional sampling
}
\examples{
calc_samplesize_prev(p_v1 = 0.1, prob = 0.95, precision = 0.25, c_ratio = 1)

}
\author{
Shirlee Wohl, Elizabeth C. Lee, Bethany L. DiPrete, and Justin Lessler
}
\concept{VOC_prevalence}
