% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDistSim.R
\docType{data}
\name{genDistSim}
\alias{genDistSim}
\title{Simulations of the genetic distance distribution}
\format{
dataframe
}
\usage{
genDistSim
}
\description{
This data object contains the genetic distance distributions
for 168 values of \eqn{R} between 1.3 and 18. The distributions represent the
the average of 1000 simulations for each value, which can be used as a
reasonable proxy for the generation distribution for large outbreaks.
}
\examples{
data(genDistSim)

}
\author{
Shirlee Wohl, John Giles, and Justin Lessler
}
\keyword{datasets}
