% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosamp-package.R
\docType{package}
\name{phylosamp-package}
\alias{phylosamp}
\alias{phylosamp-package}
\title{phylosamp: Sample Size Calculations for Molecular and Phylogenetic Studies}
\description{
Implements novel tools for estimating sample sizes needed for phylogenetic studies, including studies focused on estimating the probability of true pathogen transmission between two cases given phylogenetic linkage and studies focused on tracking pathogen variants at a population level. Methods described in Wohl, Giles, and Lessler (2021) and in Wohl, Lee, DiPrete, and Lessler (2023).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HopkinsIDD/phylosamp}
  \item Report bugs at \url{https://github.com/HopkinsIDD/phylosamp/issues}
}

}
\author{
\strong{Maintainer}: Justin Lessler \email{jlessler@unc.edu}

Authors:
\itemize{
  \item Shirlee Wohl \email{swohl@scripps.edu} [contributor]
  \item Elizabeth C Lee \email{elizabeth.c.lee@jhu.edu} [contributor]
  \item Lucy D'Agostino McGowan \email{lucydagostino@gmail.com} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID}) [contributor]
  \item John R Giles \email{jrgiles@uw.edu} [contributor]
}

}
\keyword{internal}
