% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosamp_funcs.R
\name{prob_trans_mtml}
\alias{prob_trans_mtml}
\title{Probability of transmission assuming multiple-transmission and multiple-linkage}
\usage{
prob_trans_mtml(eta, chi, rho, M, R)
}
\arguments{
\item{eta}{scalar or vector giving the sensitivity of the linkage criteria}

\item{chi}{scalar or vector giving the specificity of the linkage criteria}

\item{rho}{scalar or vector giving the proportion of the final outbreak size that is sampled}

\item{M}{scalar or vector giving the number of cases sampled}

\item{R}{scalar or vector giving the effective reproductive number of the pathogen}
}
\value{
scalar or vector giving the probability of transmission between two cases given linkage by phylogeneitic criteria
}
\description{
This function calculates the probabilitiy that two cases are linked by direct transmission
given that they have been linked by phylogenetic criteria. The multiple-transmission and multiple-linkage method assumes the following:
\enumerate{
     \item Each case \eqn{i} is the infector of \code{R} cases in the population (\eqn{N}), where \code{R} is Poisson distributed
     with its mean equal to the effective reproductive number of the pathogen.
     \item Each case \eqn{i} is allowed to be linked by the linkage criteria to multiple cases \eqn{j} in the sampled population (\eqn{M}).
     }
}
\examples{
prob_trans_mtml(eta=1, chi=1, rho=0.5, M=100, R=1) # Perfect sensitivity and specificity

prob_trans_mtml(eta=0.99, chi=0.9, rho=1, M=50, R=1)

prob_trans_mtml(eta=0.99, chi=0.9, rho=1, M=50, R=1.5)

prob_trans_mtml(eta=0.99, chi=0.9, rho=0.5, M=100, R=1)

prob_trans_mtml(eta=0.99, chi=0.9, rho=0.5, M=100, R=1.5)
}
\seealso{
Other prob_trans: \code{\link{prob_trans_mtsl}},
  \code{\link{prob_trans_stsl}}
}
\author{
John Giles, Shirlee Wohl, and Justin Lessler
}
\concept{prob_trans}
