% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosamp_funcs.R
\name{sens_spec_roc}
\alias{sens_spec_roc}
\title{Wrapper function to turn output from sens_spec_calc to values to make an ROC curve}
\usage{
sens_spec_roc(
  cutoff,
  mut_rate,
  mean_gens_pdf,
  max_link_gens = 1,
  max_gens = which(mean_gens_pdf != 0)[length(which(mean_gens_pdf != 0))],
  max_dist = max_gens * qpois(0.999, mut_rate)
)
}
\arguments{
\item{cutoff}{the maximum genetic distance at which to consider cases linked}

\item{mut_rate}{mean number of mutations per generation, assumed to be poisson distributed}

\item{mean_gens_pdf}{the density distribution of the mean number of generations between cases;
the index of this vector is assumed to be the discrete distance between cases}

\item{max_link_gens}{the maximium generations of separation for linked pairs}

\item{max_gens}{the maximum number of generations to consider, defaults to the highest
number of generations in mean_gens_pdf with a non-zero probability}

\item{max_dist}{the maximum distance to calculate, defaults to max_gens * 99.9th percentile
of mut_rate poisson distribution}
}
\value{
data frame with cutoff, sensitivity, and 1-specificity
}
\description{
Wrapper function to turn output from sens_spec_calc to values to make an ROC curve
}
\seealso{
Other mutrate_functions: 
\code{\link{gen_dists}()},
\code{\link{sens_spec_calc}()}
}
\author{
Shirlee Wohl and Justin Lessler
}
\concept{mutrate_functions}
