% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translink_prob_transmit_mtsl.R
\name{translink_prob_transmit_mtsl}
\alias{translink_prob_transmit_mtsl}
\title{Calculate probability of transmission assuming multiple-transmission and single-linkage}
\usage{
translink_prob_transmit_mtsl(chi, eta, rho, M, R)
}
\arguments{
\item{chi}{scalar or vector giving the specificity of the linkage criteria}

\item{eta}{scalar or vector giving the sensitivity of the linkage criteria}

\item{rho}{scalar or vector giving the proportion of the final outbreak size that is sampled}

\item{M}{scalar or vector giving the number of cases sampled}

\item{R}{scalar or vector giving the effective reproductive number of the pathogen}
}
\value{
scalar or vector giving the probability of transmission between two cases given linkage by phylogeneitic criteria
}
\description{
This function calculates the probabilitiy that two cases are linked by direct transmission
given that they have been linked by phylogenetic criteria. The multiple-transmission and single-linkage method assumes the following:
\enumerate{
     \item Each case \eqn{i} is, on average, the infector of \code{R} cases in the population (\eqn{N})
     \item Each case \eqn{i} is allowed to be linked by the linkage criteria to only one other case \eqn{j} in the sampled population (\eqn{M}).
     }
}
\examples{
# Perfect sensitivity and specificity
translink_prob_transmit_mtsl(eta=1, chi=1, rho=0.5, M=100, R=1)

translink_prob_transmit_mtsl(eta=0.99, chi=0.9, rho=1, M=50, R=1)

translink_prob_transmit_mtsl(eta=0.99, chi=0.9, rho=0.5, M=100, R=1)

}
\seealso{
Other transmission linkage functions: 
\code{\link{translink_expected_links_obs_mtml}()},
\code{\link{translink_expected_links_obs_mtsl}()},
\code{\link{translink_expected_links_obs_stsl}()},
\code{\link{translink_expected_links_obs}()},
\code{\link{translink_expected_links_true_mtml}()},
\code{\link{translink_expected_links_true_mtsl}()},
\code{\link{translink_expected_links_true_stsl}()},
\code{\link{translink_expected_links_true}()},
\code{\link{translink_fdr}()},
\code{\link{translink_prob_transmit_mtml}()},
\code{\link{translink_prob_transmit_stsl}()},
\code{\link{translink_samplesize}()},
\code{\link{translink_tdr}()}
}
\author{
John Giles, Shirlee Wohl, and Justin Lessler
}
\concept{transmission linkage functions}
