% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosamp_funcs.R
\name{true_pairs_mtml}
\alias{true_pairs_mtml}
\title{Expected number of true transmission pairs assuming multiple-transmission and multiple-linkage}
\usage{
true_pairs_mtml(eta, rho, M, R)
}
\arguments{
\item{eta}{scalar or vector giving the sensitivity of the linkage criteria}

\item{rho}{scalar or vector giving the proportion of the final outbreak size that is sampled}

\item{M}{scalar or vector giving the number of cases sampled}

\item{R}{scalar or vector giving the effective reproductive number of the pathogen}
}
\value{
scalar or vector giving the expected number of true transmission pairs in the sample
}
\description{
This function calculates the expected number of true transmission pairs in a sample of size \code{M}.
The multiple-transmission and multiple-linkage method assumes the following:
\enumerate{
     \item Each case \eqn{i} is the infector of \code{R} cases in the population (\eqn{N}), where \code{R} is Poisson distributed
     with its mean equal to the effective reproductive number of the pathogen.
     \item Each case \eqn{i} is allowed to be linked by the linkage criteria to multiple cases \eqn{j} in the sampled population (\eqn{M}).
     }
}
\examples{
true_pairs_mtml(eta=0.95, rho=0.2, M=1000, R=3)
}
\seealso{
Other true_pairs: \code{\link{true_pairs_mtsl}},
  \code{\link{true_pairs_stsl}}
}
\author{
John Giles, Shirlee Wohl and Justin Lessler
}
\concept{true_pairs}
