% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfreq_cdf_logistic.R
\name{varfreq_cdf_logistic}
\alias{varfreq_cdf_logistic}
\title{Function to calculate cumulative observed variant prevalence at time t given logistic growth (CDF)}
\usage{
varfreq_cdf_logistic(t, p0, r, c_ratio = 1)
}
\arguments{
\item{t}{time step number (e.g., days) at which to calculate prevalence}

\item{p0}{initial variant prevalence (# introductions / infected population size)}

\item{r}{logistic growth rate}

\item{c_ratio}{coefficient of detection ratio, calculated as the ratio of the coefficients of variant 1 to variant 2; default = 1 (no bias)}
}
\value{
scalar giving the cdf of variant prevalence at time t
}
\description{
Function to calculate cumulative observed variant prevalence at time t given logistic growth (CDF)
}
\examples{
varfreq_cdf_logistic(t = 30, p0 = 1/10000, r = 0.1, c_ratio = 1)

}
\seealso{
Other variant frequency functions: 
\code{\link{varfreq_expected_mbias}()},
\code{\link{varfreq_freq_logistic}()},
\code{\link{varfreq_obs_freq}()}
}
\author{
Shirlee Wohl, Elizabeth C. Lee, Bethany L. DiPrete, and Justin Lessler
}
\concept{variant frequency functions}
