% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartrack_prob_detect.R
\name{vartrack_prob_detect}
\alias{vartrack_prob_detect}
\title{Function to calculate the probability of detecting a variant given a sample size}
\usage{
vartrack_prob_detect(
  n,
  t = NA,
  p_v1 = NA,
  omega,
  p0 = NA,
  r = NA,
  c_ratio = 1,
  sampling_freq
)
}
\arguments{
\item{n}{sample size (either of cross-section or per timestep)}

\item{t}{time step number (e.g., days) at which variant should be detected by}

\item{p_v1}{the variant prevalence at detection}

\item{omega}{the sequencing success rate}

\item{p0}{initial variant prevalence (# introductions / infected population size)}

\item{r}{logistic growth rate}

\item{c_ratio}{coefficient of detection ratio, calculated as the ratio of the coefficients of variant 1 to variant 2. default = 1 (no bias)}

\item{sampling_freq}{the sampling frequency (must be either "xsect" or "cont")}
}
\value{
scalar of detection probability
}
\description{
Function to calculate the probability of detecting a variant given a sample size
}
\examples{
# Cross-sectional sampling
vartrack_prob_detect(p_v1 = 0.02, n = 100, omega = 0.8, c_ratio = 1, sampling_freq = "xsect")

# Periodic sampling
vartrack_prob_detect(n = 158, t = 30, omega = 0.8, p0 = 1/10000, r = 0.1, c_ratio = 1, sampling_freq = "cont")

}
\seealso{
Other variant tracking functions: 
\code{\link{vartrack_cod_ratio}()},
\code{\link{vartrack_prob_detect_cont}()},
\code{\link{vartrack_prob_detect_xsect}()},
\code{\link{vartrack_prob_prev_xsect}()},
\code{\link{vartrack_prob_prev}()},
\code{\link{vartrack_samplesize_detect_cont}()},
\code{\link{vartrack_samplesize_detect_xsect}()},
\code{\link{vartrack_samplesize_detect}()},
\code{\link{vartrack_samplesize_prev_xsect}()},
\code{\link{vartrack_samplesize_prev}()}
}
\author{
Shirlee Wohl, Elizabeth C. Lee, Bethany L. DiPrete, and Justin Lessler
}
\concept{variant tracking functions}
