% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosamp_funcs.R
\name{samplesize}
\alias{samplesize}
\title{Sample size required to obtain a defined false discovery rate given the final outbreak size}
\usage{
samplesize(eta, chi, N, R, phi, min_pairs = 1)
}
\arguments{
\item{eta}{scalar or vector giving the sensitivity of the linkage criteria}

\item{chi}{scalar or vector giving the specificity of the linkage criteria}

\item{N}{scalar or vector giving the final outbreak size}

\item{R}{scalar or vector giving the effective reproductive number of the pathogen}

\item{phi}{scalar or vector giving the desired true discovery rate (1-false discovery rate)}

\item{min_pairs}{minimum number of linked pairs observed in the sample, defaults to 1 pair (2 samples); this is to ensure reasonable results are obtained}
}
\value{
scalar or vector giving the sample size needed to meet the given conditions
}
\description{
This function calculates the sample size needed to obtain a defined false disovery rate given a final outbreak size \eqn{N}.
This function is only provided for the multiple-transmission and multiple-linkage method, which assumes the following:
\enumerate{
     \item Each case \eqn{i} is, on average, the infector of \code{R} cases in the population (\eqn{N})
     \item Each case \eqn{i} is allowed to be linked by the linkage criteria to multiple cases \eqn{j} in the sampled population (\eqn{M}).
     \item Linkage events are independent of one another (i.e, linkage of case \eqn{i} to case \eqn{j} has no bearing on linkage of case \eqn{i} to any other sample).
     }
}
\author{
John Giles, Shirlee Wohl, and Justin Lessler
}
\concept{inverse_functions}
