% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translink_samplesize.R
\name{translink_samplesize}
\alias{translink_samplesize}
\title{Calculate sample size needed to identify true transmission links}
\usage{
translink_samplesize(
  eta,
  chi,
  N,
  R = NULL,
  phi,
  min_pairs = 1,
  assumption = "mtml"
)
}
\arguments{
\item{eta}{scalar or vector giving the sensitivity of the linkage criteria}

\item{chi}{scalar or vector giving the specificity of the linkage criteria}

\item{N}{scalar or vector giving the final outbreak size}

\item{R}{scalar or vector giving the effective reproductive number of the pathogen}

\item{phi}{scalar or vector giving the desired true discovery rate (1-false discovery rate)}

\item{min_pairs}{minimum number of linked pairs observed in the sample, defaults to 1 pair (2 samples); this is to ensure reasonable results are obtained}

\item{assumption}{a character vector indicating which assumptions about transmission and linkage criteria. Default = \code{'mtml'}. Accepted arguments are:
\enumerate{
     \item \code{'stsl'} for the single-transmission single-linkage assumption.
     \item \code{'mtsl'} for the multiple-transmission single-linkage assumption.
     \item \code{'mtml'} for the multiple-transmission multiple-linkage assumption.
     }}
}
\value{
scalar or vector giving the sample size needed to meet the given conditions
}
\description{
This function calculates the sample size needed to identify transmission links at 
a pre-defined false disovery rate, given a final outbreak size \eqn{N}.
}
\examples{
translink_samplesize(eta=0.99, chi=0.995, N=100, R=1, phi=0.75)

}
\seealso{
Other transmission linkage functions: 
\code{\link{translink_expected_links_obs_mtml}()},
\code{\link{translink_expected_links_obs_mtsl}()},
\code{\link{translink_expected_links_obs_stsl}()},
\code{\link{translink_expected_links_obs}()},
\code{\link{translink_expected_links_true_mtml}()},
\code{\link{translink_expected_links_true_mtsl}()},
\code{\link{translink_expected_links_true_stsl}()},
\code{\link{translink_expected_links_true}()},
\code{\link{translink_fdr}()},
\code{\link{translink_prob_transmit_mtml}()},
\code{\link{translink_prob_transmit_mtsl}()},
\code{\link{translink_prob_transmit_stsl}()},
\code{\link{translink_tdr}()}
}
\author{
John Giles, Shirlee Wohl, and Justin Lessler
}
\concept{transmission linkage functions}
