% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartrack_prob_detect_cont.R
\name{vartrack_prob_detect_cont}
\alias{vartrack_prob_detect_cont}
\title{Function to calculate probability of detecting a variant given a per-timestep sample size assuming periodic sampling}
\usage{
vartrack_prob_detect_cont(n, t = NA, p_v1 = NA, omega, p0, r, c_ratio = 1)
}
\arguments{
\item{n}{per-timestep (e.g., per day) sample size}

\item{t}{time step number (e.g., days) at which variant should be detected by}

\item{p_v1}{the desired prevalence to detect a variant by}

\item{omega}{the sequencing success rate}

\item{p0}{initial variant prevalence (# introductions / infected population size)}

\item{r}{logistic growth rate}

\item{c_ratio}{coefficient of detection ratio, calculated as the ratio of the coefficients of variant 1 to variant 2. default = 1 (no bias)}
}
\value{
scalar of detection probability
}
\description{
Function to calculate probability of detecting a variant given a per-timestep sample size assuming periodic sampling
}
\examples{
vartrack_prob_detect_cont(n = 158, t = 30, omega = 0.8, p0 = 1/10000, r = 0.1, c_ratio = 1)

}
\seealso{
Other variant tracking functions: 
\code{\link{vartrack_cod_ratio}()},
\code{\link{vartrack_prob_detect_xsect}()},
\code{\link{vartrack_prob_detect}()},
\code{\link{vartrack_prob_prev_xsect}()},
\code{\link{vartrack_prob_prev}()},
\code{\link{vartrack_samplesize_detect_cont}()},
\code{\link{vartrack_samplesize_detect_xsect}()},
\code{\link{vartrack_samplesize_detect}()},
\code{\link{vartrack_samplesize_prev_xsect}()},
\code{\link{vartrack_samplesize_prev}()}
}
\author{
Shirlee Wohl, Elizabeth C. Lee, Bethany L. DiPrete, and Justin Lessler
}
\concept{variant tracking functions}
