% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartrack_samplesize_detect_cont.R
\name{vartrack_samplesize_detect_cont}
\alias{vartrack_samplesize_detect_cont}
\title{Function to calculate sample size needed for variant detection assuming periodic sampling}
\usage{
vartrack_samplesize_detect_cont(
  prob,
  t = NA,
  p_v1 = NA,
  omega,
  p0,
  r,
  c_ratio = 1
)
}
\arguments{
\item{prob}{desired probability of detection}

\item{t}{time step number (e.g., days) at which variant should be detected by}

\item{p_v1}{the desired prevalence to detect a variant by}

\item{omega}{the sequencing success rate}

\item{p0}{initial variant prevalence (# introductions / infected population size)}

\item{r}{logistic growth rate}

\item{c_ratio}{coefficient of detection ratio, calculated as the ratio of the coefficients of variant 1 to variant 2. default = 1 (no bias)}
}
\value{
scalar of expected sample size
}
\description{
Function to calculate sample size needed for variant detection assuming periodic sampling
}
\examples{
vartrack_samplesize_detect_cont(prob = 0.95, t = 30, omega = 0.8, p0 = 1/10000, r = 0.1, c_ratio = 1)

}
\seealso{
Other variant tracking functions: 
\code{\link{vartrack_cod_ratio}()},
\code{\link{vartrack_prob_detect_cont}()},
\code{\link{vartrack_prob_detect_xsect}()},
\code{\link{vartrack_prob_detect}()},
\code{\link{vartrack_prob_prev_xsect}()},
\code{\link{vartrack_prob_prev}()},
\code{\link{vartrack_samplesize_detect_xsect}()},
\code{\link{vartrack_samplesize_detect}()},
\code{\link{vartrack_samplesize_prev_xsect}()},
\code{\link{vartrack_samplesize_prev}()}
}
\author{
Shirlee Wohl, Elizabeth C. Lee, Bethany L. DiPrete, and Justin Lessler
}
\concept{variant tracking functions}
